backend/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
python-multipart
PyPDF2
python-docx
pydantic

passlib[bcrypt] 
python-jose

